Remove check for multiple ids in a selector
We removed this because it contained an unhandled edge case:

If your selector was something like "#foo#baz.bar" then it would correctly parse it as only one id, "#foo#bar". But if your selector was "#foo.bar#baz", it would parse two separate ids instead of one.

First, it seems that "foo#bar" is an invalid id in the first place, so it doesn't make sense to parse it as a single id. Second, "#foo#baz.bar" and "#foo.bar#baz" are equivalent (if broken) selectors, so it doesn't make sense to parse them differently.

We basically punted on this and removed the inconsistency instead of coming up with an actual fix.